Fix IsEnumDefined type mismatch preventing deep dungeon detection
IClientState.TerritoryType is ushort on TC's Dalamud API, but
ETerritoryType has a uint underlying type. IsEnumDefined requires an
exact type match, so it threw on every framework tick and
IsInDeepDungeon() always returned false, blocking the overlay from
ever drawing.

+++ Pal.Client/Floors/TerritoryState.cs
@@ -23,7 +23,7 @@ namespace Pal.Client.Floors
         public bool IsInDeepDungeon() =>
             _clientState.IsLoggedIn
             && _condition[ConditionFlag.InDeepDungeon]
-            && typeof(ETerritoryType).IsEnumDefined(_clientState.TerritoryType);
+            && typeof(ETerritoryType).IsEnumDefined((uint)_clientState.TerritoryType);
 
     }
